script/requirements.txt
pyyaml==5.4
pdfkit==0.6.1
markdown==3.1.1
